fix(workers): extract ProductionAllocationRow to its own file (Refs #2692 S6)
production_panel.dart was 732 lines (> 700 game_widgets_file_size limit).
Move the private _ProductionAllocationRow class plus its 4 allocation-row
icon constants into production_allocation_row.dart, exporting it as the
public ProductionAllocationRow widget. Production panel drops to 540 lines;
new row file is 201 lines. No behavior change.
